Move methods that interact with Repository class from Menu to DataService.
