+++ Services/Menu.cs
@@ -13,8 +13,6 @@ namespace ADOnetSakilaKoppling.Services
 {
     internal class Menu
     {
-        private const int ActorsPerColumn = 4;
-        private const int FilmsPerColumn = 3;
         private bool _running;
         private readonly Input _input;
         private readonly Output _output;
@@ -54,47 +52,6 @@ namespace ADOnetSakilaKoppling.Services
                 _menuOptions.Where(mo => mo.Id == menuChoice).First<MenuOption>().Execute();
             else
                 ShowUnexpectedInput();
-        }
-        public void PrintFilmographiesByFirstName()
-        {
-            string firstName = _input.GetString(MenuHelper.PromptFirstName);
-            List<Actor> actors = _repository.GetActorsByFirstName(firstName);
-            PrintFilmographies(actors);
-        }
-        public void PrintFilmographiesByLastName()
-        {
-            string lastName = _input.GetString(MenuHelper.PromptLastName);
-            List<Actor> actors = _repository.GetActorsByLastName(lastName);
-            PrintFilmographies(actors);
-        }
-        public void PrintFilmographiesByFullName()
-        {
-            string firstName = _input.GetString(MenuHelper.PromptFirstName);
-            string lastName = _input.GetString(MenuHelper.PromptLastName);
-            List<Actor> actors = _repository.GetActorsByFullName(firstName, lastName);
-            PrintFilmographies(actors);
-        }
-        public void PrintFilmographies(List<Actor> actors)
-        {
-            if (actors.Count > 0)
-                foreach (Actor actor in actors)
-                    PrintFilmography(actor);
-            else
-                _output.WriteWarning(MenuHelper.WarningNoActorsFound);
-            _output.ConfirmContinue();
-        }
-        private void PrintFilmography(Actor actor)
-        {
-            _output.WriteSubtitle(MenuHelper.SubtitleFilmsWithActor(actor));
-            int columnWidth = _repository.LongestFilmTitle() + 1;
-            MenuHelper.PrintList(_output, actor.Films, FilmsPerColumn, columnWidth);
-        }
-        public void PrintAllActorNames()
-        {
-            _output.WriteSubtitle(MenuHelper.SubtitleListAllActors);
-            int columnWidth = _repository.LongestActorName() + 1;
-            MenuHelper.PrintList(_output, _repository.GetAllActors(), ActorsPerColumn, columnWidth);
-            _output.ConfirmContinue();
         }
         public void ShowUnexpectedInput()
         {
