+++ Services/ProgramLauncher.cs
@@ -15,9 +15,10 @@ namespace ADOnetSakilaKoppling.Services
             Output output = new Output();
             Input input = new Input(output);
             Repository repository = new Repository();
+            DataService dataService = new DataService(input, output, repository);
             Menu menu = new Menu(input, output, repository);
 
-            MenuBuilder.BuildMenuOptions(menu);
+            MenuBuilder.BuildMenuOptions(menu, dataService);
 
             menu.Start();
         }
